requirements.txt
PyPDF2==3.0.1
<<<<<<< HEAD
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
pylint==2.17.5
=======
>>>>>>> 55d00ac82a1de281f183048895dc62a6db3a5f2a
